requirements.txt
absl-py==2.2.2
aioice==0.10.1
aiortc==1.11.0
albucore==0.0.23
albumentations==2.0.5
altair==5.5.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
av==14.3.0
blinker==1.9.0
boto3==1.38.0
botocore==1.38.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
dnspython==2.7.0
docker-pycreds==0.4.0
easydict==1.13
executing==2.2.0
facenet-pytorch==2.5.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
insightface @ git+https://github.com/nizhib/pytorch-insightface@b353a341f6f4b074581c8b1a3a8f9ed058cdcee1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mozuma==0.9.0
mozuma-clip==1.0.1
mpmath==1.3.0
namex==0.0.9
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prettytable==3.16.0
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
s3transfer==0.12.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==79.0.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
stack-data==0.6.3
streamlit==1.45.0
streamlit-webrtc==0.62.4
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xgboost==3.0.0
yacs==0.1.8
